Move unit tests as well
