+++ tests/unit/utils/mcp/test_mcp_auth_headers.py
@@ -2,7 +2,7 @@
 
 from pathlib import Path
 
-from utils.mcp_auth_headers import resolve_authorization_headers
+from utils.mcp.mcp_auth_headers import resolve_authorization_headers
 
 
 def test_resolve_authorization_headers_empty() -> None:
